Changed license headers and package name
All changes as requested in the PR to merge into the sample repo.
Maven wrapper was completely removed, as it is not strictly
necessary for the sample. It has to be user-provided now.
