+++ .gitignore
@@ -1,3 +1,6 @@
+.mvnw/**
+mvnw
+mvnw.cmd
 
 .idea/**
