+++ pom.xml
@@ -5,8 +5,14 @@
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.amazon</groupId>
-  <artifactId>appconfig-quarkus</artifactId>
+  <artifactId>sample-appconfig-quarkus</artifactId>
   <version>1.0-SNAPSHOT</version>
+  <licenses>
+    <license>
+      <name>MIT-0</name>
+    </license>
+  </licenses>
+
   <properties>
     <compiler-plugin.version>3.8.1</compiler-plugin.version>
     <maven.compiler.parameters>true</maven.compiler.parameters>
