+++ src/main/java/com/amazon/sample/appconfig/quarkus/LoanCalculationRequest.java
@@ -15,7 +15,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.amazon;
+package com.amazon.sample.appconfig.quarkus;
 
 public class LoanCalculationRequest {
     private int durationInYears;
