+++ src/main/resources/META-INF/services/org.eclipse.microprofile.config.spi.ConfigSource
@@ -1,1 +1,1 @@
-com.amazon.config.source.AwsAppConfigSource
+com.amazon.sample.appconfig.quarkus.config.source.AwsAppConfigSource
